requirements.txt
discord.py
py-cord
python-dotenv
pymongo 